refactor: update AdminLayout component filename to PascalCase
