+++ web/pages/_app.tsx
@@ -30,7 +30,7 @@ import { Router, useRouter } from 'next/router';
 
 import { RecoilRoot } from 'recoil';
 import { useEffect } from 'react';
-import AdminLayout from '../components/layouts/admin-layout';
+import AdminLayout from '../components/layouts/AdminLayout';
 import SimpleLayout from '../components/layouts/SimpleLayout';
 
 function App({ Component, pageProps }: AppProps) {
